pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "watchlist-signal-bot"
version = "0.1.0"
description = "Rule-based watchlist signal bot for KR and US symbols."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "finance-datareader>=0.9.94",
  "Jinja2>=3.1.6",
  "PyYAML>=6.0.2",
  "numpy>=2.2.4",
  "pandas>=2.2.3",
  "pyarrow>=19.0.1",
  "pykrx>=1.0.51",
  "requests>=2.32.3",
  "yfinance>=0.2.55",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/watchlist_signal_bot"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
